integration-examples/zoom-webinar-pipedrive-deals/requirements.txt
python-dotenv>=1.0.0
-r ../../core/surfe/requirements.txt
-r ../../core/integrations/zoom/requirements.txt
-r ../../core/integrations/pipedrive/requirements.txt 